transfer-execute - call directly from proxy

+++ contracts/feature-tests/async/vault/src/lib.rs
@@ -49,8 +49,9 @@ pub trait Vault {
 			&token_payment,
 			token_nonce,
 		);
-		
-		self.call_counts(b"accept_funds_echo_payment").update(|c| *c += 1);
+
+		self.call_counts(b"accept_funds_echo_payment")
+			.update(|c| *c += 1);
 
 		Ok((
 			token_identifier,
@@ -116,5 +117,5 @@ pub trait Vault {
 	/// this additional counter has the role of showing that storage also gets saved correctly.
 	#[view]
 	#[storage_mapper("call_counts")]
-	fn call_counts(&self, endpoint: &[u8]) -> SingleValueMapper<Self::Storage, u32>;
+	fn call_counts(&self, endpoint: &[u8]) -> SingleValueMapper<Self::Storage, usize>;
 }
